setDefaultPerspectiveData()
